+++ herd-server/bin/module/collection/debian_metasploit.js
@@ -1,0 +1,130 @@
+'use strict';
+
+const LinuxModule = require('../base/rdhd-mod-base_linux_module');
+
+// moduleCode: debian_airgeddon
+class DebianMetasploit extends LinuxModule
+{
+    // ************************************************************
+    //  Metasploit module
+    // ************************************************************
+    // https://www.kali.org/docs/tools/starting-metasploit-framework-in-kali/
+    //
+    // operation       :   [start/stop]
+    // ************************************************************
+
+    constructor(asset, context, session, wsServer, token)
+    {
+        super(asset, "debian_metasploit", session, wsServer, token);
+
+        //  POST parameters
+        // ******************************
+        this.auth = { username: context.username, password: context.password };
+        this.operation = context.operation;
+
+        //  Local fields
+        // ******************************
+        this.url = "https://127.0.0.1:3000/api/assets/" + this.asset.id + "/service" + this.urlReadyToken;
+    }
+
+    configure(whatIf = false)
+    {
+        let task = "sudo apt update; \
+                    sudo apt install -y build-essential zlib1g zlib1g-dev libpq-dev libpcap-dev libsqlite3-dev ruby ruby-dev && \
+                    cd /opt && \
+		    sudo git clone https://github.com/rapid7/metasploit-framework.git && \
+		    sudo chown -R `whoami` /opt/metasploit-framework && \
+                    cd metasploit-framework && \
+                    sudo gem install bundler && \
+                    sudo bundle install && \
+                    sudo cp -s /opt/metasploit-framework/msfconsole /usr/bin/msfconsole";
+
+        //  Async execution
+        // ******************************
+        this.do(task, "cmd_res", false, whatIf);
+    }
+
+    interact(whatIf = false)
+    {
+        let axios = require('axios');
+        let https = require("https");
+        let deasync = require('deasync');
+        let response = false;
+
+        if (this.validate())
+        {
+            let agent = new https.Agent({ rejectUnauthorized: false });
+            let op;
+
+            if (this.operation.toUpperCase() == "START")
+            {
+                let task = "msfconsole";
+                op = "enable";
+
+                axios.post(this.url,
+                { 
+                    operation: op,
+                    service: {
+                        type: 'TERMINAL',
+                        params: {
+                            command: task
+                        }
+                    }
+                },
+                {
+                    httpsAgent: agent,
+                    auth: this.auth
+                })
+                .then((res) => {
+                    response = res.data.data.service;
+                })
+                .catch((err) => {
+                    response = err.message;
+                });
+                deasync.loopWhile(() => !response);
+            }
+            else
+            {
+                op = "disable";
+
+                axios.post(this.url,
+                { 
+                    operation: op,
+                    service: {
+                        type: 'TERMINAL'
+                    }
+                },
+                {
+                    httpsAgent: agent,
+                    auth: this.auth
+                })
+                .then((res) => {
+                    response = res.data.data.service;
+                })
+                .catch((err) => {
+                    response = err.message;
+                });
+                deasync.loopWhile(() => !response);
+
+                this.reportAndExit(this.buildInfoMessage("Operation started"));
+            }
+        }
+        else
+        {
+            this.reportAndExit("Invalid input provided");
+        }
+        return response;
+    }
+
+    validate()
+    {
+        let result = false;
+        if ((this.operation.toUpperCase() == "START") || (this.operation.toUpperCase() == "STOP"))
+        {
+            result = true;
+        }
+        return result;
+    }
+}
+
+module.exports = DebianMetasploit